requirements.txt
networkx>=2.5
pytest>=6.2
Pillow>=8.2.0
tk==0.1.0
numpy==1.20.3
